Add custom EEPROM region.
There already exists custom_long and custom_float lines for using the BSB-LAN controller for our own custom functionality. What was lacking is the ability to save persistent data to the EEPROM in the custom functionality.

To aid in this, an 80 byte region has been added to the EEPROM which is reserved for use by custom logic. The size was chosen arbitrarily. There are 20 custom lines, so I figured 20 custom EEPROM words?