Allow changing custom EEPROM region from web interface
Because the region is entirely custom, there's no way of setting it in a user specific way in the web interface unless we add some sort of hook in for passing off to custom code.
As a workaround, instead we can make it a simple list of bytes in the web interface. The contents can be set as a comma separated list of 8-bit decimal or hexadecimal integers.